Create signal_di-lepton_L.py
Previously, I was working on separate files for each kinematic variable but have switched over to Jack's more condensed script. I have added a couple more variables (delta R and acollinearity). 

I have not committed the acollinearity equation because I wanted to check with you first. I put something like acoplanarity as a placeholder for now.  

There seems to be an error running this code where it creates pdf's that can't be opened.